requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
spacy==3.7.2
SpeechRecognition==3.10.0
gTTS==2.4.0
python-dotenv==1.0.0
